requirements.txt
Flask
pandas
requests
beautifulsoup4
mysql-connector-python
selenium
webdriver-manager
plotly
dash
bokeh